+++ index.html
@@ -8,6 +8,7 @@
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <link href="js/functions.js">
   <link href="js/main.js">
+  <script type="module" src="js/main.js"></script>
   <title>Кекстаграм</title>
 </head>
